Yangwon

+++ src/main/resources/static/js/category.js
@@ -1,3 +1,5 @@
+import validation from "./validation.js";
+
 var main = {
     init: function () {
         let _this = this;
@@ -5,7 +7,7 @@ var main = {
         let categoryDeleteBtn = document.getElementById('category-delete');
 
         categorySaveBtn.addEventListener('click', () => {
-            _this.checkDuplicate();
+            _this.check();
             // _this.save();
         });
 
@@ -14,11 +16,19 @@ var main = {
         })
     },
 
-    checkDuplicate: function () {
-        let parentId = $('#category-select option:selected').val();
+    check: function () {
+        const parentId = document.getElementById("category-select").options
+            [document.getElementById("category-select").selectedIndex].value;
+
         let name = document.getElementById('name').value;
         let selectValid = document.getElementById('category-select-valid');
         let nameValid = document.getElementById('category-name-valid');
+        const selectedCategory = document.getElementById('category-select')
+
+        selectedCategory.classList.remove('input-error-border');
+
+        validation.removeValidation(selectValid);
+        validation.removeValidation(nameValid);
 
         if (parentId !== '' && name !== '') {
             axios({
@@ -34,7 +44,7 @@ var main = {
                 if (check === false) {
                     this.save();
                 } else {
-                    alert('중복된 카테고리입니다.');
+                    alert("이미 등록된 카테고리입니다.");
                 }
             }).catch((e) => {
                 console.error(e);
@@ -42,17 +52,15 @@ var main = {
         }
 
         if (parentId === '') {
-            document.getElementById('category-select').classList.add('input-error-border');
-            selectValid.innerText = '상위 카테고리를 선택해주세요.'
-            selectValid.classList.add('input-error');
-            selectValid.style.display = 'block';
+            selectedCategory.classList.add('input-error-border');
+
+            validation.addValidation(selectValid, "상위 카테고리를 선택해주세요");
         }
 
         if (name === '') {
-            document.getElementById('name').classList.add('input-error');
-            nameValid.innerText = '카테고리명을 입력해주세요.'
-            nameValid.classList.add('input-error');
-            nameValid.style.display = 'block';
+            document.getElementById("name").classList.add("valid-custom");
+
+            validation.addValidation(nameValid, "카테고리명을 입력해주세요.");
         }
 
     },
@@ -89,15 +97,24 @@ var main = {
 
         console.log(categoryId);
 
-        axios({
-            method: "delete",
-            url: "/category/" + categoryId,
-        }).then((resp) => {
-            alert('카테고리 삭제가 완료됐습니다.');
-            window.location.reload();
-        }).catch((error) => {
-            console.log(error);
-        })
+        let selectValid = document.getElementById('category-select-valid');
+
+        if (categoryId === '') {
+            document.getElementById('category-select').classList.add('input-error-border');
+            validation.addValidation(selectValid, "삭제할 카테고리를 선택해주세요.");
+        } else {
+            validation.removeValidation(selectValid);
+
+            axios({
+                method: "delete",
+                url: "/category/" + categoryId,
+            }).then((resp) => {
+                alert('카테고리 삭제가 완료됐습니다.');
+                window.location.reload();
+            }).catch((error) => {
+                console.log(error);
+            })
+        }
     },
 }
